requirements.txt
python-dotenv
flet
SQLAlchemy
google-genai